+++ tests/unit_tests/bittensor_tests/test_receptor.py
@@ -3,7 +3,6 @@ import torch
 import bittensor
 
 from unittest.mock import MagicMock
-import bittensor.serialization as serialization
 
 wallet =  bittensor.wallet(
     path = '/tmp/pytest',
@@ -60,7 +59,7 @@ def test_receptor_neuron_request_empty():
 def test_receptor_neuron_mock_server():
     y = torch.rand(3, 3, bittensor.__network_dim__)
     
-    serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+    serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
     y_serialized = serializer.serialize(y, modality = bittensor.proto.Modality.TENSOR, from_type = bittensor.proto.TensorType.TORCH)
             
     mock_return_val = bittensor.proto.TensorMessage(
@@ -81,7 +80,7 @@ def test_receptor_neuron_mock_server():
 def test_receptor_neuron_serve_timeout():
     y = torch.rand(3, 3, bittensor.__network_dim__)
     
-    serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+    serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
     y_serialized = serializer.serialize(y, modality = bittensor.proto.Modality.TENSOR, from_type = bittensor.proto.TensorType.TORCH)
             
     mock_return_val = bittensor.proto.TensorMessage(
@@ -135,7 +134,7 @@ def test_receptor_neuron_mock_server_deserialization_error():
 def test_receptor_neuron_mock_server_shape_error():
     y = torch.rand(1, 3, bittensor.__network_dim__)
 
-    serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+    serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
     y_serialized = serializer.serialize(y, modality = bittensor.proto.Modality.TENSOR, from_type = bittensor.proto.TensorType.TORCH)
    
     mock_return_val = bittensor.proto.TensorMessage(
@@ -158,7 +157,7 @@ def test_receptor_neuron_server_response_with_nans():
     y = torch.rand(3, 3, bittensor.__network_dim__)
     y[0][0][0] = np.nan
 
-    serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+    serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
     y_serialized = serializer.serialize(y, modality = bittensor.proto.Modality.TENSOR, from_type = bittensor.proto.TensorType.TORCH)
    
     mock_return_val = bittensor.proto.TensorMessage(
