+++ tests/unit_tests/bittensor_tests/test_serialization.py
@@ -1,13 +1,8 @@
 
-import torchvision.transforms as transforms
 import torch
 import unittest
-import torchvision
 import pytest
 import bittensor
-import bittensor.serialization as serialization
-
-from random import randrange
 
 class TestSerialization(unittest.TestCase):
     config = None
@@ -18,7 +13,7 @@ class TestSerialization(unittest.TestCase):
     def test_serialize(self):
         for _ in range(10):
             tensor_a = torch.rand([12, 23])
-            serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+            serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
             content = serializer.serialize(tensor_a, modality = bittensor.proto.Modality.TENSOR, from_type = bittensor.proto.TensorType.TORCH)
             tensor_b = serializer.deserialize(content, to_type = bittensor.proto.TensorType.TORCH)
             torch.all(torch.eq(tensor_a, tensor_b))
@@ -27,17 +22,17 @@ class TestSerialization(unittest.TestCase):
         # Let's grab a random image, and try and de-serialize it incorrectly.
         image = torch.ones( [1, 28, 28] )
 
-        serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
-        with pytest.raises(serialization.SerializationTypeNotImplementedException):
+        serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+        with pytest.raises(bittensor.serializer.SerializationTypeNotImplementedException):
             serializer.serialize(image, modality = bittensor.proto.Modality.IMAGE, from_type = 11)
 
     def test_deserialization_object_type_exception(self):
         data = torch.rand([12, 23])
         
-        serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+        serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
         tensor_message = serializer.serialize(data, modality = bittensor.proto.Modality.TEXT, from_type = bittensor.proto.TensorType.TORCH)
 
-        with pytest.raises(serialization.SerializationTypeNotImplementedException):
+        with pytest.raises(bittensor.serializer.SerializationTypeNotImplementedException):
             serializer.deserialize(tensor_message, to_type = 11)
     
     def test_serialize_deserialize_image(self):
@@ -45,7 +40,7 @@ class TestSerialization(unittest.TestCase):
         # Let's grab a random image, and give it a crazy type to break the system
         image = torch.ones( [1, 28, 28] )
 
-        serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+        serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
         serialized_image_tensor_message = serializer.serialize(image, modality = bittensor.proto.Modality.IMAGE, from_type = bittensor.proto.TensorType.TORCH)
         
         assert image.requires_grad == serialized_image_tensor_message.requires_grad
@@ -56,7 +51,7 @@ class TestSerialization(unittest.TestCase):
         deserialized_image_tensor_message = serializer.deserialize(serialized_image_tensor_message, to_type = bittensor.proto.TensorType.TORCH)
         assert serialized_image_tensor_message.requires_grad == deserialized_image_tensor_message.requires_grad
         assert serialized_image_tensor_message.shape == list(deserialized_image_tensor_message.shape)
-        assert serialization.torch_dtype_to_bittensor_dtype(deserialized_image_tensor_message.dtype) != bittensor.proto.DataType.UNKNOWN
+        assert bittensor.serializer.torch_dtype_to_bittensor_dtype(deserialized_image_tensor_message.dtype) != bittensor.proto.DataType.UNKNOWN
 
         assert torch.all(torch.eq(deserialized_image_tensor_message, image))
 
@@ -74,7 +69,7 @@ class TestSerialization(unittest.TestCase):
         for i, ts in enumerate(ts_list):
             data[i, 0:ts.size()[0]] = ts
 
-        serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+        serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
         serialized_data_tensor_message = serializer.serialize(data, modality = bittensor.proto.Modality.TEXT, from_type = bittensor.proto.TensorType.TORCH)
        
         assert data.requires_grad == serialized_data_tensor_message.requires_grad
@@ -85,7 +80,7 @@ class TestSerialization(unittest.TestCase):
         deserialized_data_tensor_message = serializer.deserialize(serialized_data_tensor_message, to_type = bittensor.proto.TensorType.TORCH)
         assert serialized_data_tensor_message.requires_grad == deserialized_data_tensor_message.requires_grad
         assert serialized_data_tensor_message.shape == list(deserialized_data_tensor_message.shape)
-        assert serialization.torch_dtype_to_bittensor_dtype(deserialized_data_tensor_message.dtype) != bittensor.proto.DataType.UNKNOWN
+        assert bittensor.serializer.torch_dtype_to_bittensor_dtype(deserialized_data_tensor_message.dtype) != bittensor.proto.DataType.UNKNOWN
 
         assert torch.all(torch.eq(deserialized_data_tensor_message, data))
 
@@ -93,7 +88,7 @@ class TestSerialization(unittest.TestCase):
     def test_serialize_deserialize_tensor(self):
         data = torch.rand([12, 23])
 
-        serializer = serialization.get_serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
+        serializer = bittensor.serializer( serialzer_type = bittensor.proto.Serializer.MSGPACK )
         serialized_tensor_message = serializer.serialize(data, modality = bittensor.proto.Modality.TENSOR, from_type = bittensor.proto.TensorType.TORCH)
        
         assert data.requires_grad == serialized_tensor_message.requires_grad
@@ -104,11 +99,11 @@ class TestSerialization(unittest.TestCase):
         deserialized_tensor_message = serializer.deserialize(serialized_tensor_message, to_type = bittensor.proto.TensorType.TORCH)
         assert serialized_tensor_message.requires_grad == deserialized_tensor_message.requires_grad
         assert serialized_tensor_message.shape == list(deserialized_tensor_message.shape)
-        assert serialization.torch_dtype_to_bittensor_dtype(deserialized_tensor_message.dtype) == bittensor.proto.DataType.FLOAT32
+        assert bittensor.serializer.torch_dtype_to_bittensor_dtype(deserialized_tensor_message.dtype) == bittensor.proto.DataType.FLOAT32
 
         assert torch.all(torch.eq(deserialized_tensor_message, data))
 
     
     def test_bittensor_dtype_to_torch_dtype(self):
-        with pytest.raises(serialization.DeserializationException):
-            serialization.bittensor_dtype_to_torch_dtype(11)
+        with pytest.raises(bittensor.serializer.DeserializationException):
+            bittensor.serializer.bittensor_dtype_to_torch_dtype(11)
